requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiohttp>=3.10.0
feedparser>=6.0.11
pyyaml>=6.0.2
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.0
